refactor: enhance Makefile for file exclusion, improve README for clarity on backend execution, and update syntax documentation for community recruitment

+++ pkg/compute/kernel/cuda/backend.go
@@ -9,9 +9,9 @@ int cuda_device_count();
 void cleanup_cuda_pools();
 
 typedef struct {
-    uint32_t kind;
-    uint32_t start;
-    uint32_t span;
+    uint64_t kind;
+    uint64_t start;
+    uint64_t span;
     uint64_t threshold;
 } predicate_device_spec_t;
